# swagger.yaml
openapi: 3.0.3
info:
  title: API de Promessas de Campanha
  version: 1.0.0
  description: |
    API para cadastro e consulta de candidatos e suas promessas de campanha.
servers:
  - url: http://localhost:3000
paths:
...
  /api/v1/auth/register:
    post:
      summary: Registra um novo usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterInput'
      responses:
        '201':
          description: Usuário criado e token emitido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email já cadastrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/login:
    post:
      summary: Autentica um usuário existente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        '200':
          description: Autenticação bem-sucedida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/candidates:
    get:
      summary: Lista todos os candidatos
      parameters:
        - in: query
          name: electionId
          schema:
            type: integer
          description: Filtra candidatos por eleição
        - in: query
          name: officeId
          required: true
          schema:
            type: integer
          description: Filtra candidatos pelo cargo (obrigatório)
        - in: query
          name: state_code
          required: false
          schema:
            type: integer
          description: Filtra candidatos por estado (opcional; deve condizer com o cargo)
        - in: query
          name: city_id
          required: false
          schema:
            type: integer
          description: Filtra candidatos por cidade (opcional; deve condizer com o cargo)
      responses:
        '200':
          description: Lista ordenada de candidatos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Candidate'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Cria um novo candidato
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCandidateInput'
      responses:
        '201':
          description: Candidato criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Candidate'
        '400':
          description: Requisição inválida (campos obrigatórios ausentes)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Token ausente ou inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/candidates/{candidateId}/promises:
    get:
      summary: Lista promessas de um candidato
      parameters:
        - in: path
          name: candidateId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de promessas do candidato
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Promise'
        '400':
          description: ID inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Candidato não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Cria uma nova promessa para o candidato
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: candidateId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePromiseInput'
      responses:
        '201':
          description: Promessa criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Promise'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Candidato não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Token ausente ou inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/elections:
    get:
      summary: Lista eleições cadastradas
      responses:
        '200':
          description: Lista de eleições
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Election'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Cria uma nova eleição
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateElectionInput'
      responses:
        '201':
          description: Eleição criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Election'
        '400':
          description: Dados inválidos ou ano duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Token ausente ou inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Maria
        email:
          type: string
          format: email
          example: maria@example.com
      required:
        - id
        - name
        - email
    RegisterInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 6
      required:
        - name
        - email
        - password
    LoginInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOi...
        user:
          $ref: '#/components/schemas/User'
      required:
        - token
        - user
    Candidate:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Maria Silva
        political_party_id:
          type: integer
          nullable: true
          example: 99
        political_party:
          type: string
          nullable: true
          example: MDB
        office_id:
          type: integer
          example: 5
        state_code:
          type: integer
          nullable: true
          example: 35
        city_id:
          type: integer
          nullable: true
          example: 3550308
        city_name:
          type: string
          nullable: true
          example: Sao Paulo
        promises_count:
          type: integer
          example: 3
        comments_count:
          type: integer
          example: 5
        election_id:
          type: integer
          nullable: true
          example: 10
        election_year:
          type: integer
          nullable: true
          example: 2024
        office:
          type: string
          example: Prefeita
        state_name:
          type: string
          nullable: true
          example: Sao Paulo
        created_at:
          type: string
          format: date-time
          example: '2024-01-01T00:00:00.000Z'
        updated_at:
          type: string
          format: date-time
          example: '2024-01-02T00:00:00.000Z'
      required:
        - id
        - name
        - office_id
        - office
        - created_at
        - updated_at
    Promise:
      type: object
      properties:
        id:
          type: integer
        candidate_id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/PromiseStatus'
        progress:
          type: integer
          minimum: 0
          maximum: 100
        comments_count:
          type: integer
          example: 3
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        comments:
          type: array
          items:
            $ref: '#/components/schemas/PromiseComment'
      required:
        - id
        - candidate_id
        - title
        - status
        - progress
        - created_at
        - updated_at
        - comments
    PromiseComment:
      type: object
      properties:
        id:
          type: integer
        promise_id:
          type: integer
        content:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - promise_id
        - content
        - created_at
    CreateCandidateInput:
      type: object
      properties:
        name:
          type: string
          example: João Souza
        political_party_id:
          type: integer
          nullable: true
          example: 99
        election_id:
          type: integer
          nullable: true
          example: 10
        office_id:
          type: integer
          example: 5
        state_code:
          type: integer
          nullable: true
          example: 35
          description: |
            Obrigatório para Governador, Senador, Deputado Federal e Deputado Estadual.
            Proibido para Presidente e cargos municipais (Prefeito/Vereador).
        city_id:
          type: integer
          nullable: true
          example: 3550308
          description: |
            Obrigatório para cargos municipais (Prefeito/Vereador) e deve corresponder ao estado da cidade.
            Proibido para Presidente e cargos estaduais/federais.
      required:
        - name
        - office_id
    CreatePromiseInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/PromiseStatus'
        progress:
          type: integer
          minimum: 0
          maximum: 100
      required:
        - title
...
    Election:
      type: object
      properties:
        id:
          type: integer
        year:
          type: integer
        description:
          type: string
          nullable: true
        type:
          $ref: '#/components/schemas/ElectionType'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - year
        - type
        - created_at
        - updated_at
    CreateElectionInput:
      type: object
      properties:
        year:
          type: integer
        description:
          type: string
          nullable: true
      required:
        - year
...
    PromiseStatus:
      type: string
      enum:
        - NOT_STARTED
        - IN_PROGRESS
        - COMPLETED
    ElectionType:
      type: string
      description: Tipo de eleição/cargo
      enum:
        - FEDERAL_ESTADUAL
        - MUNICIPAL
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Erro interno do servidor
      required:
        - error
  responses:
    InternalError:
      description: Erro interno inesperado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'